chore: #306 코드리뷰 반영

+++ Spoony-iOS/Spoony-iOS/Network/Model/ExploreModel/FilteredFeedRequest.swift
@@ -13,6 +13,6 @@ struct FilteredFeedRequest: Codable {
     let regionIds: [Int]
     let ageGroups: [String]
     let sortBy: String
-    let cursor: Int64
-    let size: Int32
+    let cursor: Int
+    let size: Int
 }
